refactor: group Flame map render helpers under render submodule (#3878)
Move GP ownership tint, warp-zone edge geometry, and resource disc palette
into features/game/flame/render/ with a public barrel export; update imports,
generator output path, and SPEC references.

Refs #3878
