+++ SPEC/program/map-region-map-render.md
@@ -13,7 +13,7 @@
 | `region_map_component_render_political.dart` | `_CtRegionMapRenderPolitical` | GP ownership tint, province/faction borders, province names, hovered-province glow |
 | `region_map_component_render_markers.dart` | `_CtRegionMapRenderMarkers` | Capitals, towns, warp zones, civilians, fleets, hover selector, tile selection highlights |
 
-Shared constants and small pure helpers used across layers: `region_map_component_shared.dart`. Pure warp-glow segment math (test without canvas): `warp_zone_edge_geometry.dart`.
+Shared constants and small pure helpers used across layers: `region_map_component_shared.dart`. Pure warp-glow segment math (test without canvas): `render/warp_zone_edge_geometry.dart`.
 
 ## Paint order
