+++ SPEC/program/repo-and-packages.md
@@ -27,7 +27,7 @@ Monorepo layout:
 - **`features/game/flame/controls/`** — in-game map shell chrome: top/tab bar ([GameMapControls](../../app/lib/features/game/flame/controls/game_map_controls.dart)), corner tool row ([GameMapCornerControls](../../app/lib/features/game/flame/controls/game_map_corner_controls.dart)), and empire left rail ([GameMapEmpireLeftRail](../../app/lib/features/game/flame/controls/game_map_empire_left_rail.dart)). Public barrel: `controls/controls.dart`.
 - **`features/game/flame/overlays/`** — map and session overlays/dialogs: province detail hosts ([GameMapProvinceDetailSidePanel](../../app/lib/features/game/flame/overlays/game_map_province_detail_side_panel.dart), [GameMapNarrowDetailOverlaySlot](../../app/lib/features/game/flame/overlays/game_map_narrow_detail_overlay.dart)), turn confirmation/processing dialogs, victory overlay, exit confirm, and debug console panel. Public barrel: `overlays/overlays.dart`.
 - **`features/game/flame/minimap/`** — [GameRegionMinimap](../../app/lib/features/game/flame/minimap/game_region_minimap.dart) widget and viewport math helpers. Public barrel: `minimap/minimap.dart`.
-- **Import boundary (`repo.app_flame_map_import_boundary`):** `map_area/` and `region_map/` must not import `map_state/`; `map_state/` may import `map_area` public exports only (`map_area/map_area.dart`). Legacy re-export shims under `flame/` root are forbidden — import submodule barrels directly.
+- **`features/game/flame/render/`** — pure map render helpers: GP ownership tint layer, warp-zone edge geometry, and resource extraction disc palette table. Public barrel: `render/render.dart`. `map_area/` and `region_map/` must not import `map_state/`; `map_state/` may import `map_area` public exports only (`map_area/map_area.dart`). Legacy re-export shims under `flame/` root are forbidden — import submodule barrels directly.
 - **Widgetbook catalogs** (`widgetbook_host/lib/catalogs/`) — dev-only Widgetbook stories and catalog `part` files. `app/lib/widgetbook.dart` is a thin re-export shim for local `flutter run -t lib/widgetbook.dart`; production analyze scope for `app/lib/features`, `app/lib/core`, and `app/lib/widgets` excludes relocated catalogs (`repo.app_widgetbook_catalog_location`).
 
 ---
