+++ SPEC/ui/pixel-art-ui-catalog.md
@@ -167,7 +167,7 @@ tokens") and §S4 ("Zero `const Color(0x...)` in `app/lib/features/` and
 | `app/lib/features/game/flame/region_map_component_render_political.dart` | Same as above. |
 | `app/lib/features/game/flame/region_map_component_render_markers.dart` | Same as above. |
 | `app/lib/features/game/flame/minimap/game_region_minimap.dart` | Same as above (minimap canvas). |
-| `app/lib/features/game/flame/resource_icon_disc_palette.dart` | Pixel-art palette data (per-resource hue lookup table; generated by `tool/generate_resource_icon_disc_palette.dart`). |
+| `app/lib/features/game/flame/render/resource_icon_disc_palette.dart` | Pixel-art palette data (per-resource hue lookup table; generated by `tool/generate_resource_icon_disc_palette.dart`). |
 | `app/lib/features/debug_log/debug_log_viewer_screen.dart` | Dev-tooling screen SYS10001 — relaxed per #2914 Risks / edge cases. |
 | `app/lib/features/game/flame/overlays/debug_console_overlay_panel.dart` | Dev-tooling screen SYS20001 — relaxed per #2914 Risks / edge cases. |
 | `app/lib/features/game/widgets/chrome/**` | Ct-* catalog widget implementations. These widgets implement the design-system primitives; consumer code in the rest of `features/**` must use the catalog widget, not raw Material colors. |
