+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -10,7 +10,7 @@ import 'package:colonizethis_map/colonizethis_map.dart';
 import 'package:flame/components.dart';
 import 'package:flutter/material.dart';
 
-import '../gp_ownership_tint_layer.dart';
+import '../render/gp_ownership_tint_layer.dart';
 import '../caches/civilian_icon_cache.dart';
 import '../caches/fleet_icon_cache.dart';
 import 'region_map_boundary_visibility.dart';
@@ -19,7 +19,7 @@ import '../caches/resource_icon_cache.dart';
 import '../caches/province_label_icon_cache.dart';
 import '../tilesets/tilesets.dart';
 import '../caches/town_icon_cache.dart';
-import '../warp_zone_edge_geometry.dart';
+import '../render/warp_zone_edge_geometry.dart';
