+++ app/lib/features/game/flame/render/render.dart
@@ -1,0 +1,6 @@
+/// Pure map render helpers and palette tables (Refs #3878 Phase 3).
+library;
+
+export 'gp_ownership_tint_layer.dart';
+export 'resource_icon_disc_palette.dart';
+export 'warp_zone_edge_geometry.dart';
